% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galdraw.R
\name{galdraw}
\alias{galdraw}
\title{Draw the Jupiter and Galilean satellites}
\usage{
galdraw(year, month, day, hour, minute)
}
\arguments{
\item{year}{Type in the year (integer number greater then or equal 0).}

\item{month}{Type in the month (integer number from 1 to 12).}

\item{day}{Type in the day (integer number from 1 to 31).}

\item{hour}{Type in the hour (integer number from 0 to 23).}

\item{minute}{Type in the minute (integer number from 0 to 59).}
}
\value{
A plot.
}
\description{
\code{galdraw()} is used to show the positions of the four greatest satellites
of Jupiter (called Galilean satellites). Positions are shown for any
given time (ET – Ephemeris Time) with respect to the planet, as seen from the Earth.
}
\examples{
galdraw(2025, 10, 22, 23, 30)
}
