% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galilean_satellites.R
\name{galsat}
\alias{galsat}
\title{Calculates the positions of the Galilean satellites for given time (ET – Ephemeris Time).}
\usage{
galsat(year, month, day, hour, minute)
}
\arguments{
\item{year}{Type in the year (integer number greater then 0).}

\item{month}{Type in the month (integer number from 1 to 12).}

\item{day}{Type in the day (integer number from 1 to 31).}

\item{hour}{Type in the hour (integer number from 0 to 23).}

\item{minute}{Type in the minute (integer number from 0 to 59).}
}
\value{
A data.frame
}
\description{
Calculates the positions of the Galilean satellites for given time (ET – Ephemeris Time).
}
\examples{
galsat(2025, 10, 22, 23, 30)
}
